ETL/requirements.txt
annotated-types==0.6.0
debugpy==1.8.1
et-xmlfile==1.1.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
typing_extensions==4.10.0
tzdata==2024.1
